rename remaining .svg.tsx files in app
