+++ components/admin_console/billing/billing_summary/billing_summary.tsx
@@ -10,7 +10,7 @@ import FormattedMarkdownMessage from 'components/formatted_markdown_message';
 import BlockableLink from 'components/admin_console/blockable_link';
 import OverlayTrigger from 'components/overlay_trigger';
 import Tooltip from 'components/tooltip';
-import UpgradeSvg from 'components/common/svg_images_components/upgrade.svg';
+import UpgradeSvg from 'components/common/svg_images_components/upgrade_svg';
 import EmptyBillingHistorySvg from 'components/common/svg_images_components/empty_billing_history_svg';
 
 import {trackEvent} from 'actions/telemetry_actions';
