+++ components/common/svg_images_components/background_svg.tsx
