+++ components/common/svg_images_components/cloud_trial_svg.tsx
