+++ components/common/svg_images_components/mattermost_cloud_svg.tsx
