+++ components/common/svg_images_components/private_cloud_svg.tsx
