+++ components/common/svg_images_components/upgrade_svg.tsx
