+++ components/next_steps_view/images/download-apps-svg.tsx
