+++ components/purchase_modal/purchase_modal.tsx
@@ -30,9 +30,9 @@ import Badge from 'components/widgets/badges/badge';
 import OverlayTrigger from 'components/overlay_trigger';
 import Tooltip from 'components/tooltip';
 import LoadingSpinner from 'components/widgets/loading/loading_spinner';
-import UpgradeSvg from 'components/common/svg_images_components/upgrade.svg';
-import BackgroundSvg from 'components/common/svg_images_components/background.svg';
-import MattermostCloudSvg from 'components/common/svg_images_components/mattermost_cloud.svg';
+import UpgradeSvg from 'components/common/svg_images_components/upgrade_svg';
+import BackgroundSvg from 'components/common/svg_images_components/background_svg';
+import MattermostCloudSvg from 'components/common/svg_images_components/mattermost_cloud_svg';
 
 import {getNextBillingDate} from 'utils/utils';
